Added artifical intelligence Goals. Added base class Actor to Player and Monster. Added PathfindAgent as detour agent wrapper. Added number of argument checks to ScriptManager. Typedefed Ogre::Vector3 to Vector3 and Ogre::Quaternion to Quaternion. Added getScene in BasicComponent. Added some groups in Group.h.

+++ NH2012/Engine/BoidExtension.cpp
@@ -23,7 +23,7 @@ void BoidExtension::update(double elapsedSeconds)
 
   if(node)
   {
-    node->setPosition(position.x, position.y, position.z);
+    node->setPosition((float) position.x, (float) position.y, (float) position.z);
   }
 }
